remove da layer and da implementations
