+++ block/block-manager.md
@@ -53,7 +53,7 @@ config|config.BlockManagerConfig|block manager configurations (see config option
 genesis|*cmtypes.GenesisDoc|initialize the block manager with genesis state (genesis configuration defined in `config/genesis.json` file under the app directory)
 store|store.Store|local datastore for storing rollup blocks and states (default local store path is `$db_dir/rollkit` and `db_dir` specified in the `config.toml` file under the app directory)
 mempool, proxyapp, eventbus|mempool.Mempool, proxy.AppConnConsensus, *cmtypes.EventBus|for initializing the executor (state transition function). mempool is also used in the manager to check for availability of transactions for lazy block production
-dalc|da.DataAvailabilityLayerClient|the data availability light client used to submit and retrieve blocks to DA network
+dalc|da.DAClient|the data availability light client used to submit and retrieve blocks to DA network
 blockstore|*goheaderstore.Store[*types.Block]|to retrieve blocks gossiped over the P2P network
 
 Block manager configuration options:
