+++ block/manager.go
@@ -66,8 +66,7 @@ type Manager struct {
 
 	executor *state.BlockExecutor
 
-	dalc      da.DataAvailabilityLayerClient
-	retriever da.BlockRetriever
+	dalc *da.DAClient
 	// daHeight is the height of the latest processed DA block
 	daHeight uint64
 
@@ -112,7 +111,7 @@ func NewManager(
 	store store.Store,
 	mempool mempool.Mempool,
 	proxyApp proxy.AppConnConsensus,
-	dalc da.DataAvailabilityLayerClient,
+	dalc *da.DAClient,
 	eventBus *cmtypes.EventBus,
 	logger log.Logger,
 	blockStore *goheaderstore.Store[*types.Block],
@@ -168,7 +167,6 @@ func NewManager(
 		store:       store,
 		executor:    exec,
 		dalc:        dalc,
-		retriever:   dalc.(da.BlockRetriever), // TODO(tzdybal): do it in more gentle way (after MVP)
 		daHeight:    s.DAHeight,
 		// channels are buffered to avoid blocking on input/output operations, buffer sizes are arbitrary
 		HeaderCh:          make(chan *types.SignedHeader, channelLength),
@@ -197,9 +195,8 @@ func getAddress(key crypto.PrivKey) ([]byte, error) {
 }
 
 // SetDALC is used to set DataAvailabilityLayerClient used by Manager.
-func (m *Manager) SetDALC(dalc da.DataAvailabilityLayerClient) {
+func (m *Manager) SetDALC(dalc *da.DAClient) {
 	m.dalc = dalc
-	m.retriever = dalc.(da.BlockRetriever)
 }
 
 // GetStoreHeight returns the manager's store height
@@ -530,7 +527,7 @@ func (m *Manager) processNextDABlock(ctx context.Context) error {
 
 func (m *Manager) fetchBlock(ctx context.Context, daHeight uint64) (da.ResultRetrieveBlocks, error) {
 	var err error
-	blockRes := m.retriever.RetrieveBlocks(ctx, daHeight)
+	blockRes := m.dalc.RetrieveBlocks(ctx, daHeight)
 	switch blockRes.Code {
 	case da.StatusError:
 		err = fmt.Errorf("failed to retrieve block: %s", blockRes.Message)
