+++ block/manager_test.go
@@ -12,9 +12,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	goDATest "github.com/rollkit/go-da/test"
 	"github.com/rollkit/rollkit/config"
 	"github.com/rollkit/rollkit/da"
-	mockda "github.com/rollkit/rollkit/da/mock"
 	"github.com/rollkit/rollkit/store"
 	test "github.com/rollkit/rollkit/test/log"
 	"github.com/rollkit/rollkit/types"
@@ -96,10 +96,8 @@ func TestInitialState(t *testing.T) {
 	}
 }
 
-func getMockDALC(logger log.Logger) da.DataAvailabilityLayerClient {
-	dalc := &mockda.DataAvailabilityLayerClient{}
-	_ = dalc.Init([8]byte{}, nil, nil, logger)
-	_ = dalc.Start()
+func getMockDALC(logger log.Logger) *da.DAClient {
+	dalc := &da.DAClient{DA: goDATest.NewDummyDA(), Logger: log.TestingLogger()}
 	return dalc
 }
