+++ config/config_test.go
@@ -47,7 +47,6 @@ func TestViperAndCobra(t *testing.T) {
 	assert.NoError(v.BindPFlags(cmd.Flags()))
 
 	assert.NoError(cmd.Flags().Set(flagAggregator, "true"))
-	assert.NoError(cmd.Flags().Set(flagDALayer, "foobar"))
 	assert.NoError(cmd.Flags().Set(flagDAConfig, `{"json":true}`))
 	assert.NoError(cmd.Flags().Set(flagBlockTime, "1234s"))
 	assert.NoError(cmd.Flags().Set(flagNamespaceID, "0102030405060708"))
@@ -56,7 +55,6 @@ func TestViperAndCobra(t *testing.T) {
 	assert.NoError(nc.GetViperConfig(v))
 
 	assert.Equal(true, nc.Aggregator)
-	assert.Equal("foobar", nc.DALayer)
 	assert.Equal(`{"json":true}`, nc.DAConfig)
 	assert.Equal(1234*time.Second, nc.BlockTime)
 	assert.Equal(types.NamespaceID{1, 2, 3, 4, 5, 6, 7, 8}, nc.NamespaceID)
