+++ node/full_client_test.go
@@ -26,8 +26,9 @@ import (
 	"github.com/libp2p/go-libp2p/core/crypto"
 	"github.com/libp2p/go-libp2p/core/peer"
 
+	goDATest "github.com/rollkit/go-da/test"
 	"github.com/rollkit/rollkit/config"
-	mockda "github.com/rollkit/rollkit/da/mock"
+	"github.com/rollkit/rollkit/da"
 	"github.com/rollkit/rollkit/store"
 	test "github.com/rollkit/rollkit/test/log"
 	"github.com/rollkit/rollkit/test/mocks"
@@ -79,7 +80,7 @@ func getRPC(t *testing.T) (*mocks.Application, *FullClient) {
 	key, _, _ := crypto.GenerateEd25519Key(crand.Reader)
 	signingKey, _, _ := crypto.GenerateEd25519Key(crand.Reader)
 	ctx := context.Background()
-	node, err := newFullNode(ctx, config.NodeConfig{DALayer: "newda"}, key, signingKey, proxy.NewLocalClientCreator(app), &cmtypes.GenesisDoc{ChainID: "test"}, test.NewFileLogger(t))
+	node, err := newFullNode(ctx, config.NodeConfig{}, key, signingKey, proxy.NewLocalClientCreator(app), &cmtypes.GenesisDoc{ChainID: "test"}, test.NewFileLogger(t))
 	require.NoError(err)
 	require.NotNil(node)
 
@@ -177,7 +178,7 @@ func TestGenesisChunked(t *testing.T) {
 	mockApp.On(InitChain, mock.Anything).Return(abci.ResponseInitChain{})
 	privKey, _, _ := crypto.GenerateEd25519Key(crand.Reader)
 	signingKey, _, _ := crypto.GenerateEd25519Key(crand.Reader)
-	n, _ := newFullNode(context.Background(), config.NodeConfig{DALayer: "newda"}, privKey, signingKey, proxy.NewLocalClientCreator(mockApp), genDoc, test.NewFileLogger(t))
+	n, _ := newFullNode(context.Background(), config.NodeConfig{}, privKey, signingKey, proxy.NewLocalClientCreator(mockApp), genDoc, test.NewFileLogger(t))
 
 	rpc := NewFullClient(n)
 
@@ -486,7 +487,6 @@ func TestTx(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	node, err := newFullNode(ctx, config.NodeConfig{
-		DALayer:    "newda",
 		Aggregator: true,
 		BlockManagerConfig: config.BlockManagerConfig{
 			BlockTime: 1 * time.Second, // blocks must be at least 1 sec apart for adjacent headers to get verified correctly
@@ -726,7 +726,7 @@ func createGenesisValidators(t *testing.T, numNodes int, appCreator func(require
 		wg.Add(1)
 	}
 
-	dalc := &mockda.DataAvailabilityLayerClient{}
+	dalc := &da.DAClient{DA: goDATest.NewDummyDA()}
 	ds, err := store.NewDefaultInMemoryKVStore()
 	require.Nil(err)
 	err = dalc.Init([8]byte{}, nil, ds, test.NewFileLogger(t))
@@ -747,7 +747,6 @@ func createGenesisValidators(t *testing.T, numNodes int, appCreator func(require
 		nodes[i], err = newFullNode(
 			ctx,
 			config.NodeConfig{
-				DALayer:    "newda",
 				Aggregator: true,
 				BlockManagerConfig: config.BlockManagerConfig{
 					BlockTime:   1 * time.Second,
@@ -915,7 +914,6 @@ func TestMempool2Nodes(t *testing.T) {
 	// make node1 an aggregator, so that node2 can start gracefully
 	node1, err := newFullNode(ctx, config.NodeConfig{
 		Aggregator: true,
-		DALayer:    "newda",
 		P2P: config.P2PConfig{
 			ListenAddress: "/ip4/127.0.0.1/tcp/9001",
 		},
@@ -927,7 +925,6 @@ func TestMempool2Nodes(t *testing.T) {
 	require.NotNil(node1)
 
 	node2, err := newFullNode(ctx, config.NodeConfig{
-		DALayer: "newda",
 		P2P: config.P2PConfig{
 			ListenAddress: "/ip4/127.0.0.1/tcp/9002",
 			Seeds:         "/ip4/127.0.0.1/tcp/9001/p2p/" + id1.Pretty(),
@@ -1010,7 +1007,6 @@ func TestStatus(t *testing.T) {
 	node, err := newFullNode(
 		context.Background(),
 		config.NodeConfig{
-			DALayer: "newda",
 			P2P: config.P2PConfig{
 				ListenAddress: "/ip4/0.0.0.0/tcp/26656",
 			},
@@ -1124,7 +1120,6 @@ func TestFutureGenesisTime(t *testing.T) {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 	node, err := newFullNode(ctx, config.NodeConfig{
-		DALayer:    "newda",
 		Aggregator: true,
 		BlockManagerConfig: config.BlockManagerConfig{
 			BlockTime: 200 * time.Millisecond,
